Verify aggregation pushdown isn't affected by column name length limits
In 5e520fcea34149c223090b51f781e93f73165a76 we introduced changes which
prevent a long synthetic column name to be pushed down into remote
database during join pushdown leading to query failures because for
example Oracle 12 and older have a small limit of 30 characters.

The only other pushdown operation which adds synthetic columns in the
pushed down part is aggregation pushdown. However it's not practically
affected since the synthetic aggregation columns are always of the form
`_pfgnrtd_ + nextSyntheticColumnId` which means it's limited to 9 + 10
characters which is small enough to be supported by all databases.