Respect Hive statistics_enabled session in Glue metastore
Before the change, the Hive Glue metastore wiring would respect the
`hive.table-statistics-enabled` configuration property, but not the
`statistics_enabled` session property.

After the change, handling of stats for Hive Glue metastore is same as
for other metastores. That means in particular that some of the
`DefaultGlueColumnStatisticsProvider` code will execute implicitly, as
not all stats-related operations are guarded by `statistics_enabled`
toggle today (see e.g. `SemiTransactionalHiveMetastore` and implicit
stats updates after table creation or alteration). This is considered
not a blocker.