Limit partitionDataSizeSinceLastRebalance to 0
Since we estimate the partitionDataSize based on
partitionRowCount and total data processed. It is
possible that the estimated partitionDataSize is
slightly less than it was estimated at the last
rebalance cycle. That's because for a given partition,
row count hasn't increased, however overall data
processed has increased. Therefore, we need to
make sure that the estimated partitionDataSize
since last rebalance is always greater than 0.
Otherwise, it will affect the ordering of
minTaskBuckets priority queue.