Properly finish stage on noMorePartitions signal
Theoretically it is possible that all task partitions for a stage
complete before we get noMorePartitions signal from SplitAssigner.
In such case we should execute same finish routie as we do when last
task partition completes and noMorePartitions is already set.

This is cleanup, not a bugfix as currently no SplitAssigner
(even though it is not guaranteed by interface) would seal
all of the task partitions without returning noMorePartitions flag in
the same AssignmentResult object.