Do not fail stage after already done
Ignore failure within stage execution if it already done.
This is to work around the problem which was introduced with
https://github.com/trinodb/trino/commit/39b04491c05.
With that change we close EventDrivenTaskSource when stage completes,
and as a result of that it may emit an failure event due to internall
processes being cancelled.
Handling of event in `StageExecution.fail()` did close a number of
objects needed for query completion (see createStageExecutionCloser()).
As a result the query could hang.