Fix table listing concurrency in FileHiveMetastore
Before the change, `FileHiveMetastore` would rely on
`fileSystem.listFiles` to find schemas or tables within a schema.
`listFiles` is recursive and, on a local file system, fails when
files/directories are being modified concurrently. This commit replaces
recursive `listFiles` listing with non-recursive `listDirectories`
leveraging the fact that listed entities are represented as directories.