Bump number of queryExecutor threads in tests
While by default number of queryExecutor threads is 1000 in tests we
limitted it to 10. This values was too small though. If tests are run in
parallel all threads may be consumed by long running jobs running
EventDrivenFaultTolerantQueryScheduler.Scheduler#run. In such case we
are not able to run Future callback which use same executor and are also
needed for queries to progress. If that happens test queries would hang.