Fix thread leak in MinioClient
Our `MinioClient` wraps MinIO's own `MinioClient` (`io.minio.MinioClient`).
The MinIO `MinioClient` builder creates a new `OkHttpClient` and
previously we were not closing the HTTP client, which resulted in
leaking threads (`OkHttpClient` has an executor service). There seems to
be no way to dispose of MinIO's `MinioClient`'s resources, so create the
HTTP client explicitly so that we can close it.