Replace "is distinct from" operator with "identical" operator
"is distinct from" is analogous to "not equal". Having to deal with
the negation makes it harder to reason about it and requires a
double negation to get meaning analogous to "equal" (i.e., not(is_distinct_from(x)))

We replace that operator with a new "identical" operator. It has similar
semantics to "equal", except that it considers nulls identical to each other.
I.e., it behaves like "is not distinct from".