Fix logic obtaining workers memory in FTE execution
The change to filter out coordinator from output of
ClusterMemoryManager::getWorkerMemoryInfo
(https://github.com/trinodb/trino/pull/23079) was beneficial for how
ExponentialGrowthPartitionMemoryEstimator behaves if coordinator is
differently sized than workers.

Yet it introduced regression as same method was used in
BinPackingNodeAllocatorService and as a result COORDINATOR_ONLY tasks
could not be scheduled at all if node-scheduler.include-coordinator was
set to false.

Unfortunately we had node-scheduler.include-coordinator set to true in
all the tests so it was not catch by automation.

This commit adds a regression test too.